==> analyze/collect_metrics/analyze_crate.io/01_code_document/02_grouping_same_files.ipynb <==
# Importing libraries

import os, glob, csv
from os import path
import pandas as pd

# Defining directories

userhome = os.path.expanduser('~')
analyze_dir = userhome + r"/different-diff/analyze/collect_metrics/analyze_crate.io/"

# Load diff dataset

algorithms = ['histogram','minimal','myers','patience']

df_diffhist = pd.read_csv(analyze_dir + '02_diff-file/diffhist-metrics.csv')
df_diffmin = pd.read_csv(analyze_dir + '02_diff-file/diffmin-metrics.csv')
df_diffmyers = pd.read_csv(analyze_dir + '02_diff-file/diffmyers-metrics.csv')
df_diffpat = pd.read_csv(analyze_dir + '02_diff-file/diffpat-metrics.csv')

# Grouping same files and summarize the number of added and deleted lines

data_frame = [df_diffhist, df_diffmin, df_diffmyers, df_diffpat]
summarize = []
for i, df in enumerate(data_frame):
    temp = df.groupby('filename')[['#insertions','#deletions','#line_changed']].sum()
    summarize.append(temp)

summarize[0]

for z in range(0,len(summarize)):
    summarize[z].to_csv(analyze_dir + r'03_summarize/' + algorithms[z] + '_file_sum.csv')

# Merging dataset

Load the summarize dataset

df_sum = []
for a, alg in enumerate(algorithms):
    df_sum.append(pd.read_csv(analyze_dir + r'03_summarize/' + alg + '_file_sum.csv'))

df_sum[0]

#merge df histogram and minimal
dm_histmin = df_sum[0].merge(df_sum[1], on=['filename'], how='outer', suffixes=('_histogram','_minimal'))
#merge df myers and patience
dm_myerspat = df_sum[2].merge(df_sum[3], on=['filename'], how='outer', suffixes=('_myers','_patience'))

#merge all
dm_all = dm_histmin.merge(dm_myerspat, on=['filename'], how='outer')

#fill NAs with 0 and sort the dataset descending
dm_all.fillna(0, inplace=True)
dm_all = dm_all.sort_values(by = '#line_changed_histogram', ascending=False).reset_index(drop=True)

cols = ['filename','#insertions_histogram','#deletions_histogram','#insertions_minimal','#deletions_minimal',
        '#insertions_myers','#deletions_myers','#insertions_patience','#deletions_patience']
dm_all = dm_all[cols]
dm_all

#saving to CSV file
dm_all.to_csv(analyze_dir + '04_merge_dataframe/datamerge_all_algorithms.csv', index=False)
